feat: LectureProgress 생성 및 갱신구현 및 검증
- lectureProgress 스키마도 재구성
- 아직 controller 연결 전이라 수강권 인가 검증 수행
- 진행률(시청시간/영상길이)은 BigDecimal이 아닌 Integer (그정도의 정밀도 필요없음)

+++ backend/src/test/java/com/team08/backend/domain/lectureprogress/repository/LectureProgressRepositoryTest.java
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -119,7 +118,7 @@ class LectureProgressRepositoryTest {
                 userId,
                 lastPositionSeconds,
                 lastPositionSeconds,
-                BigDecimal.valueOf(50.00),
+                50,
                 completed,
                 completed ? updatedAt : null,
                 updatedAt,
